move datastorage parsing to new file
